noframework/python/pyproject.toml
[project]
name = "baseline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pathspec>=0.11.0",
    "binaryornot>=0.4.4",
    "openai>=1.0.0",
    "pydantic>=2.7.0",
    "tree-sitter-languages>=1.10.2",
    "pytest>=9.0.1",
    "tree-sitter==0.20.2",
    "setuptools>=80.9.0",
    "dspy-ai>=3.0.4",
    "uvicorn>=0.38.0",
    "fastapi>=0.121.3",
    "jinja2>=3.1.6",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
]
